用户分数排名查询和添加

+++ service/src/main/java/com/example/exam/vo/AnswerRecordVo.java
@@ -18,9 +18,9 @@ public class AnswerRecordVo {
     /**
      * 正确数
      */
-    private Integer rightNum;
+    private Double rightNum;
     /**
      * 问题数
      */
-    private Integer numOfQuestion;
+    private Double numOfQuestion;
 }
